factor out component; fix error handling
